Revise function references to include argument bindings.
This will enable behaviors like lambda expressions. This seems
cleaner than having each function reference be "maybe this is just
the location of the function, or maybe it's N layers of indirection
that add arguments at each layer".

It also conveniently removes the syntactic ambiguity (in the early
stages of parsing) between a variable name and a function reference.