Get rid of the "return" keyword.
This is probably slightly controversial and it's possible I'll throw
in some other keyword as a replacement. However, it's misleading
when it comes to how this was being used within "if" blocks, and
hopefully the fact that everything else is an assignment will keep
this clear. I should revisit this decision after a syntax for
monads is determined.