Implement a lesser version of composed/extended literals.
This version crashes at runtime instead of failing validation if a
literal value doesn't satisfy a requires block. This sucks as a
developer experience, but I'll need to rewrite the compiler anyway,
and currently this would require either writing a second, lesser
interpreter to run code during validation, or allow ValidatedFunctions
to exist that aren't fully validated.