Allow the language server to handle multiple source files as a module
Currently they must all be in the same directory and have a
'module.conf' file in the directory.

The LSP recommends watching files through the client, but I haven't
been able to get this to work, so we read files from the file
system instead. Still need to implement watching for updates.