Do initial interpreter test setup, with a few tests passing
The current approach to running tests is too slow -- each test is
run as a JUnit test in Kotlin that, after doing the translation to
JSON, runs a Gradle process in order to run a bootstrapped Node
process (since the Node location depends on OS) that runs the test.

The expected future approach will be to have a task in Kotlin code
that will translate every file in the corpus to the JSON format,
then run a unit test framework on the Node/TS side to go through
and test all of them and report the results.