Remove parameter types for function expressions
These weren't being used in this incarnation of the language
(except in some buggy cases). In principle we could support these,
but we'd have to add the notion of a type parameter to a function
type and not just a function signature.

(Not supporting this does prevent us from, say, having a reduce()
method on a ListLike interface, which would want to allow the user
to specify the output type as in List.reduce but would come in the
form of a function binding.)